requirements.txt
#Packages for the Django Project
Django
djangorestframework
django-cors-headers
asgiref
psycopg2
sqlparse
markdown
django-filter
whitenoise
gunicorn
django-heroku
scikit-learn
pandas
django-extensions
django-bulk-update-or-create
